Protect against resolving stateless alerts
At some point, the Argus API changed in a way that caused nav-argus-glue
to post stateless NAV alerts as stateful Argus Incidents. Once this
issue was resolved, this code path would attempt to close those Argus
Incidents and fail, because the corresponding NAV alert is stateless,
and therefore has an end_time value of None. None cannot be
compared to datetime values.

This change ensures the current datetime is used when closing those
types of incidents.

Fixes #8