Add timeout and verify select return value
navargus would sometimes get stuck in a tight loop of select/read when
there was no data to read. It is not clear why the select call exited
early, but it is likely a problem that the code didn't even verify the
return value from select to ensure that it actually did exit because
there was data to read from the stream.

This adds:

- Verification that the stream was in fact readable after select exits.
- Debug logging of what's going on in the select loop.
- A timeout value to ensure we sometimes DO break out of loop.