Continue reading from stream when buffer is full
There was a bug in the stream reader that caused some events to get
lost. The reader only reads `buf_size` bytes at a time. When a
single read operation returns a block of the maximum buffer size,
there may in fact be more data waiting to be read from the stream.

However, select does not signal us that there is more data to read
at this point (since nothing new was written to the stream since the
last select call).

To make sure we read everything the stream has to offer, this code
changes the logic to read more without waiting for a select call, if
the last read operation returned a full buffer of data.